refactor out convert_to_str
