+++ src/zenml/artifact_stores/base_artifact_store.py
@@ -33,6 +33,7 @@ from pydantic import root_validator
 
 from zenml.enums import StackComponentType
 from zenml.exceptions import ArtifactStoreInterfaceError
+from zenml.io import fileio
 from zenml.stack import Flavor, StackComponent, StackComponentConfig
 from zenml.utils import io_utils
 
@@ -53,7 +54,7 @@ def _sanitize_potential_path(potential_path: Any) -> Any:
         path.
     """
     if isinstance(potential_path, bytes):
-        path = io_utils.convert_to_str(potential_path)
+        path = fileio.convert_to_str(potential_path)
     elif isinstance(potential_path, str):
         path = potential_path
     else:
