+++ tests/unit/utils/test_io_utils.py
@@ -301,13 +301,3 @@ def test_get_parent_gets_the_parent_directory(tmp_path) -> None:
     )
     parent = io_utils.get_parent(os.path.join(tmp_path, "new_dir/new_dir2"))
     assert parent == "new_dir"
-
-
-def test_convert_to_str_converts_to_string(tmp_path) -> None:
-    """Test that convert_to_str converts bytes to a string"""
-    assert isinstance(
-        io_utils.convert_to_str(bytes(str(tmp_path), "ascii")), str
-    )
-    assert io_utils.convert_to_str(bytes(str(tmp_path), "ascii")) == str(
-        tmp_path
-    )
